Simplify optics to Traversal/Optional/Lens/Prism
