+++ arrow-libs/optics/arrow-optics/src/commonMain/kotlin/arrow/optics/Copy.kt
@@ -8,14 +8,14 @@ public annotation class OpticsCopyMarker
 @OpticsCopyMarker
 public interface Copy<A> {
   /**
-   * Changes the value of the element(s) pointed by the [Setter].
+   * Changes the value of the element(s) pointed by the [Traversal].
    */
-  public infix fun <B> Setter<A, B>.set(b: B)
+  public infix fun <B> Traversal<A, B>.set(b: B)
 
   /**
    * Transforms the value of the element(s) pointed by the [Traversal].
    */
-  public infix fun <B> Setter<A, B>.transform(f: (B) -> B)
+  public infix fun <B> Traversal<A, B>.transform(f: (B) -> B)
 
   /**
    * Declares a block in which all optics are nested within
@@ -46,10 +46,10 @@ public interface Copy<A> {
 
 // mutable builder of copies
 private class CopyImpl<A>(var current: A): Copy<A> {
-  override fun <B> Setter<A, B>.set(b: B) {
+  override fun <B> Traversal<A, B>.set(b: B) {
     current = this.set(current, b)
   }
-  override fun <B> Setter<A, B>.transform(f: (B) -> B) {
+  override fun <B> Traversal<A, B>.transform(f: (B) -> B) {
     current = this.modify(current, f)
   }
 }
