+++ arrow-libs/optics/arrow-optics/src/commonMain/kotlin/arrow/optics/Lens.kt
@@ -28,9 +28,9 @@ public typealias Lens<S, A> = PLens<S, S, A, A>
  * @param A the focus of a [PLens]
  * @param B the modified focus of a [PLens]
  */
-public interface PLens<S, T, A, B> : PGetter<S, T, A>, POptional<S, T, A, B> {
+public interface PLens<S, T, A, B> : POptional<S, T, A, B> {
 
-  override fun get(source: S): A
+  public fun get(source: S): A
 
   override fun set(source: S, focus: B): T
 
@@ -86,7 +86,10 @@ public interface PLens<S, T, A, B> : PGetter<S, T, A>, POptional<S, T, A, B> {
 
   public companion object {
 
-    public fun <S> id(): PIso<S, S, S, S> = PIso.id<S>()
+    public fun <S> id(): Lens<S, S> = Lens(
+      get = { it },
+      set = { _, s -> s }
+    )
 
     /**
      * [PLens] that takes either [S] or [S] and strips the choice of [S].
@@ -210,5 +213,14 @@ public interface PLens<S, T, A, B> : PGetter<S, T, A>, POptional<S, T, A, B> {
     @JvmStatic
     public fun <A, B, C> tripleThird(): Lens<Triple<A, B, C>, C> =
       triplePThird()
+
+    /**
+     * Defines equality between String and [List] of [Char]
+     */
+    @JvmStatic
+    public fun stringToList(): Lens<String, List<Char>> = PLens(
+      get = CharSequence::toList,
+      set = { _, ss -> ss.joinToString(separator = "") }
+    )
   }
 }
