+++ arrow-libs/optics/arrow-optics/src/commonMain/kotlin/arrow/optics/typeclasses/At.kt
@@ -2,14 +2,10 @@ package arrow.optics.typeclasses
 
 import arrow.core.None
 import arrow.core.Option
-import arrow.optics.Fold
-import arrow.optics.Getter
-import arrow.optics.Iso
 import arrow.optics.Lens
 import arrow.optics.Optional
 import arrow.optics.PLens
 import arrow.optics.Prism
-import arrow.optics.Setter
 import arrow.optics.Traversal
 import kotlin.jvm.JvmStatic
 
@@ -41,16 +37,6 @@ public fun interface At<S, I, A> {
     this@at.compose(this@At.at(i))
 
   /**
-   *  DSL to compose [At] with an [Iso] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Iso] with a focus in [S]
-   * @param i index [I] to zoom into [S] and find focus [A]
-   * @return [Lens] with a focus in [A] at given index [I].
-   */
-  public fun <T> Iso<T, S>.at(i: I): Lens<T, A> =
-    this.compose(this@At.at(i))
-
-  /**
    *  DSL to compose [At] with a [Prism] for a structure [S] to focus in on [A] at given index [I].
    *
    * @receiver [Prism] with a focus in [S]
@@ -69,24 +55,6 @@ public fun interface At<S, I, A> {
   public fun <T> Optional<T, S>.at(i: I): Optional<T, A> = this.compose(this@At.at(i))
 
   /**
-   * DSL to compose [At] with a [Getter] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Getter] with a focus in [S]
-   * @param i index [I] to zoom into [S] and find focus [A]
-   * @return [Getter] with a focus in [A] at given index [I].
-   */
-  public fun <T> Getter<T, S>.at(i: I): Getter<T, A> = this.compose(this@At.at(i))
-
-  /**
-   * DSL to compose [At] with a [Setter] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Setter] with a focus in [S]
-   * @param i index [I] to zoom into [S] and find focus [A]
-   * @return [Setter] with a focus in [A] at given index [I].
-   */
-  public fun <T> Setter<T, S>.at(i: I): Setter<T, A> = this.compose(this@At.at(i))
-
-  /**
    * DSL to compose [At] with a [Traversal] for a structure [S] to focus in on [A] at given index [I].
    *
    * @receiver [Traversal] with a focus in [S]
@@ -95,26 +63,7 @@ public fun interface At<S, I, A> {
    */
   public fun <T> Traversal<T, S>.at(i: I): Traversal<T, A> = this.compose(this@At.at(i))
 
-  /**
-   * DSL to compose [At] with a [Fold] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Fold] with a focus in [S]
-   * @param i index [I] to zoom into [S] and find focus [A]
-   * @return [Fold] with a focus in [A] at given index [I].
-   */
-  public fun <T> Fold<T, S>.at(i: I): Fold<T, A> = this.compose(this@At.at(i))
-
   public companion object {
-
-    /**
-     * Lift an instance of [At] using an [Iso].
-     *
-     * @param AT [At] that can provide [Lens] for a structure [U] with a focus in [A] with given index [I].
-     * @param iso [Iso] that defines an isomorphism between [S] and [U]
-     * @return [At] to provide [Lens] for structure [S] with focus in [A] at given index [I]
-     */
-    public fun <S, U, I, A> fromIso(AT: At<U, I, A>, iso: Iso<S, U>): At<S, I, A> =
-      At { i -> iso compose AT.at(i) }
 
     @JvmStatic
     public fun <K, V> map(): At<Map<K, V>, K, Option<V>> =
