+++ arrow-libs/optics/arrow-optics/src/commonMain/kotlin/arrow/optics/typeclasses/Index.kt
@@ -4,13 +4,11 @@ import arrow.core.NonEmptyList
 import arrow.core.left
 import arrow.core.right
 import arrow.core.toNonEmptyListOrNull
-import arrow.optics.Fold
-import arrow.optics.Iso
 import arrow.optics.Lens
 import arrow.optics.Optional
+import arrow.optics.PLens
 import arrow.optics.POptional
 import arrow.optics.Prism
-import arrow.optics.Setter
 import arrow.optics.Traversal
 import kotlin.jvm.JvmStatic
 
@@ -41,15 +39,6 @@ public fun interface Index<S, I, A> {
   public fun <T> Lens<T, S>.index(i: I): Optional<T, A> = this.compose(this@Index.index(i))
 
   /**
-   *  DSL to compose [Index] with an [Iso] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Iso] with a focus in [S]
-   * @param i index [I] to focus into [S] and find focus [A]
-   * @return [Optional] with a focus in [A] at given index [I].
-   */
-  public fun <T> Iso<T, S>.index(i: I): Optional<T, A> = this.compose(this@Index.index(i))
-
-  /**
    *  DSL to compose [Index] with a [Prism] for a structure [S] to focus in on [A] at given index [I].
    *
    * @receiver [Prism] with a focus in [S]
@@ -68,15 +57,6 @@ public fun interface Index<S, I, A> {
   public fun <T> Optional<T, S>.index(i: I): Optional<T, A> = this.compose(this@Index.index(i))
 
   /**
-   * DSL to compose [Index] with a [Setter] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Setter] with a focus in [S]
-   * @param i index [I] to focus into [S] and find focus [A]
-   * @return [Setter] with a focus in [A] at given index [I].
-   */
-  public fun <T> Setter<T, S>.index(i: I): Setter<T, A> = this.compose(this@Index.index(i))
-
-  /**
    * DSL to compose [Index] with a [Traversal] for a structure [S] to focus in on [A] at given index [I].
    *
    * @receiver [Traversal] with a focus in [S]
@@ -86,15 +66,6 @@ public fun interface Index<S, I, A> {
   public fun <T> Traversal<T, S>.index(i: I): Traversal<T, A> = this.compose(this@Index.index(i))
 
   /**
-   * DSL to compose [Index] with a [Fold] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Fold] with a focus in [S]
-   * @param i index [I] to focus into [S] and find focus [A]
-   * @return [Fold] with a focus in [A] at given index [I].
-   */
-  public fun <T> Fold<T, S>.index(i: I): Fold<T, A> = this.compose(this@Index.index(i))
-
-  /**
    * DSL to compose [Index] with a [Lens] for a structure [S] to focus in on [A] at given index [I].
    *
    * @receiver [Lens] with a focus in [S]
@@ -102,15 +73,6 @@ public fun interface Index<S, I, A> {
    * @return [Optional] with a focus in [A] at given index [I].
    */
   public operator fun <T> Lens<T, S>.get(i: I): Optional<T, A> = this.compose(this@Index.index(i))
-
-  /**
-   *  DSL to compose [Index] with an [Iso] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Iso] with a focus in [S]
-   * @param i index [I] to focus into [S] and find focus [A]
-   * @return [Optional] with a focus in [A] at given index [I].
-   */
-  public operator fun <T> Iso<T, S>.get(i: I): Optional<T, A> = this.compose(this@Index.index(i))
 
   /**
    *  DSL to compose [Index] with a [Prism] for a structure [S] to focus in on [A] at given index [I].
@@ -131,15 +93,6 @@ public fun interface Index<S, I, A> {
   public operator fun <T> Optional<T, S>.get(i: I): Optional<T, A> = this.compose(this@Index.index(i))
 
   /**
-   * DSL to compose [Index] with a [Setter] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Setter] with a focus in [S]
-   * @param i index [I] to focus into [S] and find focus [A]
-   * @return [Setter] with a focus in [A] at given index [I].
-   */
-  public operator fun <T> Setter<T, S>.get(i: I): Setter<T, A> = this.compose(this@Index.index(i))
-
-  /**
    * DSL to compose [Index] with a [Traversal] for a structure [S] to focus in on [A] at given index [I].
    *
    * @receiver [Traversal] with a focus in [S]
@@ -148,25 +101,7 @@ public fun interface Index<S, I, A> {
    */
   public operator fun <T> Traversal<T, S>.get(i: I): Traversal<T, A> = this.compose(this@Index.index(i))
 
-  /**
-   * DSL to compose [Index] with a [Fold] for a structure [S] to focus in on [A] at given index [I].
-   *
-   * @receiver [Fold] with a focus in [S]
-   * @param i index [I] to focus into [S] and find focus [A]
-   * @return [Fold] with a focus in [A] at given index [I].
-   */
-  public operator fun <T> Fold<T, S>.get(i: I): Fold<T, A> = this.compose(this@Index.index(i))
-
   public companion object {
-    /**
-     * Lift an instance of [Index] using an [Iso].
-     *
-     * @param ID [Index] instance to provide a [Optional] to focus into [S] at [I]
-     * @param iso [Iso] that defines an isomorphism between a type [S] and [A]
-     * @return [Index] for a structure [S] to focus in an optional [A] at a given index [I]
-     */
-    public fun <S, A, I, B> fromIso(ID: Index<A, I, B>, iso: Iso<S, A>): Index<S, I, B> =
-      Index { i -> iso compose ID.index(i) }
 
     /**
      * [Index] instance definition for [List].
@@ -224,7 +159,7 @@ public fun interface Index<S, I, A> {
     @JvmStatic
     public fun string(): Index<String, Int, Char> =
       Index { i ->
-        Iso.stringToList() compose Index.list<Char>().index(i)
+        PLens.stringToList() compose Index.list<Char>().index(i)
       }
   }
 }
