--- arrow-libs/optics/arrow-optics/src/commonTest/kotlin/arrow/optics/OptionalGetterTest.kt
@@ -1,24 +1,0 @@
-package arrow.optics
-
-import arrow.core.test.UnitSpec
-import io.kotest.matchers.shouldBe
-import io.kotest.property.Arb
-import io.kotest.property.arbitrary.boolean
-import io.kotest.property.arbitrary.int
-
-class OptionalGetterTest : UnitSpec() {
-
-  init {
-    "get should return value if predicate is true and null if otherwise" {
-      checkAll(Arb.int(), Arb.boolean()) { int, predicate ->
-        OptionalGetter.filter<Int> { predicate }.getOrNull(int) shouldBe (if (predicate) int else null)
-      }
-    }
-
-    "getAll should return the old list if predicate is true" {
-      checkAll(Arb.list(Arb.int()), Arb.boolean()) { list, predicate ->
-        (Fold.list<Int>() compose OptionalGetter.filter { predicate }).getAll(list) shouldBe (if (predicate) list else emptyList())
-      }
-    }
-  }
-}
