+++ arrow-libs/optics/arrow-optics/src/commonTest/kotlin/arrow/optics/TraversalTest.kt
@@ -3,7 +3,6 @@ package arrow.optics
 import arrow.core.test.UnitSpec
 import arrow.core.test.generators.functionAToB
 import arrow.core.test.generators.option
-import arrow.optics.test.laws.SetterLaws
 import arrow.optics.test.laws.TraversalLaws
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.char
@@ -22,13 +21,6 @@ class TraversalTest : UnitSpec() {
         bGen = Arb.int(),
         funcGen = Arb.functionAToB(Arb.int()),
       ),
-
-      SetterLaws.laws(
-        setter = Traversal.list(),
-        aGen = Arb.list(Arb.int()),
-        bGen = Arb.int(),
-        funcGen = Arb.functionAToB(Arb.int()),
-      )
     )
 
     testLaws(
