+++ arrow-libs/optics/arrow-optics/src/commonTest/kotlin/arrow/optics/instances/ConsInstanceTest.kt
@@ -7,7 +7,6 @@ import arrow.optics.typeclasses.Cons
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.char
 import io.kotest.property.arbitrary.int
-import io.kotest.property.arbitrary.list
 import io.kotest.property.arbitrary.pair
 import io.kotest.property.arbitrary.string
