+++ arrow-libs/optics/arrow-optics/src/commonTest/kotlin/arrow/optics/std/ListTest.kt
@@ -2,12 +2,8 @@ package arrow.optics.std
 
 import arrow.core.test.UnitSpec
 import arrow.core.test.generators.functionAToB
-import arrow.core.test.generators.option
-import arrow.optics.Iso
 import arrow.optics.Optional
-import arrow.optics.test.laws.IsoLaws
 import arrow.optics.test.laws.OptionalLaws
-import arrow.optics.test.laws.SetterLaws
 import arrow.optics.test.laws.TraversalLaws
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.int
@@ -30,12 +26,6 @@ class ListTest : UnitSpec() {
         bGen = Arb.int(),
         funcGen = Arb.functionAToB(Arb.int()),
       ),
-      SetterLaws.laws(
-        setter = Optional.listHead<Int>(),
-        aGen = Arb.list(Arb.int()),
-        bGen = Arb.int(),
-        funcGen = Arb.functionAToB(Arb.int()),
-      )
     )
 
     testLaws(
@@ -45,16 +35,6 @@ class ListTest : UnitSpec() {
         aGen = Arb.list(Arb.int()),
         bGen = Arb.list(Arb.int()),
         funcGen = Arb.functionAToB(Arb.list(Arb.int())),
-      )
-    )
-
-    testLaws(
-      "Iso list to Option Nel - ",
-      IsoLaws.laws(
-        iso = Iso.listToOptionNel(),
-        aGen = Arb.list(Arb.int()),
-        bGen = Arb.option(Arb.nonEmptyList(Arb.int())),
-        funcGen = Arb.functionAToB(Arb.option(Arb.nonEmptyList(Arb.int()))),
       )
     )
   }
